[frenemies] Data fetching improvements + UI tweaks (#8156)
