+++ .changeset/curvy-clocks-doubt.md
@@ -1,0 +1,6 @@
+---
+"@mysten/wallet-kit": patch
+"@mysten/sui.js": patch
+---
+
+Add basic formatting utilities
