+++ apps/explorer/src/ui/InternalLink.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { formatAddress, formatDigest } from '../utils/stringUtils';
+import { formatAddress, formatDigest } from '@mysten/sui.js';
 
 import { Link, type LinkProps } from '~/ui/Link';
