+++ apps/explorer/src/utils/stringUtils.ts
@@ -48,20 +48,6 @@ export function truncate(fullStr: string, strLen: number, separator?: string) {
     );
 }
 
-const ELLIPSIS = '\u{2026}';
-export function formatAddress(address: string) {
-    const offset = address.startsWith('0x') ? 2 : 0;
-
-    return `0x${address.slice(offset, offset + 4)}${ELLIPSIS}${address.slice(
-        -4
-    )}`;
-}
-
-export function formatDigest(digest: string) {
-    // Use 10 first characters
-    return `${digest.slice(0, 10)}${ELLIPSIS}`;
-}
-
 export async function extractFileType(
     displayString: string,
     signal: AbortSignal
