+++ dapps/frenemies/src/network/queries/epoch.ts
@@ -5,14 +5,14 @@ import { SystemEpochInfo, SYSTEM_EPOCH_INFO } from "../types";
 import { useQuery } from "@tanstack/react-query";
 import provider from "../provider";
 import { bcs } from "../bcs";
-import { config } from "../../config";
+import { useEffect } from "react";
 
 /**
  * Fetch the most recent SystemEpochInfo event to get the
  * data on the current epoch and timestamps.
  */
 export function useEpoch() {
-  return useQuery(
+  const ret = useQuery(
     ["epoch"],
     async (): Promise<{
       timestamp: number;
@@ -39,8 +39,27 @@ export function useEpoch() {
       };
     },
     {
-      // refetch 4 times per epoch
-      refetchInterval: (+config.VITE_EPOCH_LEN * 60000) / 4,
+      // Refetch every 10 minutes:
+      refetchInterval: 10 * 60 * 1000,
     }
   );
+
+  // This does additional refetching one minute after we think the epoch changes:
+  useEffect(() => {
+    if (!ret.data || !ret.data.timestamp || !ret.data.prevTimestamp) return;
+    const { timestamp, prevTimestamp } = ret.data;
+    const timePassed = Date.now() - timestamp;
+    const prevEpochLength = timestamp - prevTimestamp;
+    const timeLeft = prevEpochLength - timePassed;
+
+    const timeout = setTimeout(() => {
+      ret.refetch();
+    }, timeLeft + 60 * 1000);
+
+    return () => {
+      clearTimeout(timeout);
+    };
+  }, [ret.data]);
+
+  return ret;
 }
