+++ sdk/typescript/src/index.ts
@@ -26,6 +26,7 @@ export * from './signers/signer-with-provider';
 
 export * from './types';
 export * from './utils/api-endpoints';
+export * from './utils/format';
 
 export * from './framework';
